Nuevo archivo SVG para la ilustración de términos y condiciones. Implementado formulario de creacion y edición de oficios con validación y capacidad de carga de archivos. Desarrollada vista de índice de oficios con componente Livewire para manejo dinámico de datos. Agregado componente de alertas para requests via ajax notificaciones de usuario. Diseñado layout de resultados de búsqueda de oficios con renderizado condicional. Establecido layout de la página de términos y condiciones con contenido y estilos. Actualizacion completa de funcion de manejo de archivos para negocios, productos y oficios. Agregago hover.css para mejorar interaccion con usuarios Actualizacion mensajes de vacio cuando no hay negocios, productos u oficios.

+++ app/Http/Controllers/Admin/AdminBusinessController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Business;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Str;
@@ -24,7 +25,9 @@ class AdminBusinessController extends Controller
      */
     public function create()
     {
-        return view('admin.sections.business.show');
+        $business = new Business();
+        $business->id = 0;
+        return view('admin.sections.business.edit', compact('business'));
     }
 
     /**
@@ -32,7 +35,61 @@ class AdminBusinessController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        // Crear el negocio con user_id = 0
+        $business = new Business();
+        $business->fill($request->only([
+            'name',
+            'categories_id',
+            'keywords',
+            'email',
+            'email_2',
+            'description',
+            'phone',
+            'whatsapp',
+            'web',
+            'facebook',
+            'instagram',
+            'x',
+            'tiktok',
+            'mercadolibre',
+            'yapo',
+            'aliexpress',
+            'address',
+            'number',
+            'latitude',
+            'longitude',
+        ]));
+        $business->user_id = 0; // Negocio sin dueño asignado
+        $business->save();
+
+        // Carpeta para guardar imágenes
+        $folder = $business->id;
+        $filePath = public_path('uploads/business/'.$folder);
+
+        if (!file_exists($filePath)) {
+            mkdir($filePath, 0777, true);
+        }
+        //guarda las imagenes
+        if( $request->hasFile('gallery')){
+            var_dump('has gallery');
+            foreach ($request->file('gallery') as $file){
+                $ext = $file->getClientOriginalExtension();
+                $fileName = time() . '_' . Str::random(10) . '.' . $ext;
+                $file->move( $filePath , $fileName);
+            }
+        }
+
+        if( $request->hasFile('avatar')){
+            var_dump('has avatar');
+            $avatar = $request->file('avatar');
+
+            $save_path = public_path('uploads/business/'.$folder.'/_avatar.'.$avatar->extension());
+
+            $img = Image::read($avatar)->cover(500,500,'center');
+            $img->save($save_path);
+        }
+
+        return Redirect::route('admin_business.edit', $business)->with(['status' => 'success', 'message' => 'Negocio creado correctamente']);
     }
 
     /**
@@ -48,7 +105,11 @@ class AdminBusinessController extends Controller
      */
     public function edit(Business $business)
     {
-        return view('admin.sections.business.edit', ['business' => Business::findOrFail($business->id)]);
+        
+        return view('admin.sections.business.edit', [
+            'business' => Business::findOrFail($business->id),
+            'user' => User::find($business->user_id)
+        ]);
     }
 
     /**
@@ -58,6 +119,7 @@ class AdminBusinessController extends Controller
     {
         $business->fill($request->only([
             'name',
+            'user_id',
             'categories_id',
             'keywords',
             'email',
@@ -88,7 +150,8 @@ class AdminBusinessController extends Controller
 
         if( $request->hasFile('gallery')){
             foreach ($request->file('gallery') as $file){
-                $fileName = Str::random(10)  .'.'.$file->extension();
+                $ext = $file->getClientOriginalExtension();
+                $fileName = time() . '_' . Str::random(10) . '.' . $ext;
                 $file->move( $filePath , $fileName);
             }
         }
@@ -97,8 +160,9 @@ class AdminBusinessController extends Controller
             $avatar = $request->file('avatar');
 
             $old_avatar = get_images_from_folder('business',$folder,'avatar');
-            if ($old_avatar != '') {
-                $avatar_path = public_path('uploads/business/'.$folder.'/'. get_images_from_folder('business',$folder,'avatar'));
+
+            if ($old_avatar != '' && $old_avatar != 'default/_avatar.jpg') {
+                $avatar_path = public_path('uploads/business/'. get_images_from_folder('business',$folder,'avatar'));
                 unlink( $avatar_path );
             }
